requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
openai
PyPDF2
Pillow
matplotlib
seaborn
pandas